+++ Mopy/bash/game/oblivion/__init__.py
@@ -28,7 +28,6 @@
 from oblivion_const import bethDataFiles, allBethFiles
 from ... import brec
 from ...brec import *
-from ...bolt import Flags, DataDict, StateError
 
 # Util Constants ---------------------------------------------------------------
 #--Null strings (for default empty byte arrays)
@@ -1503,699 +1502,6 @@ class esp:
 #------------------------------------------------------------------------------
 from records import * # MUST BE HERE otherwise all hell breaks loose
 #------------------------------------------------------------------------------
-class MreQust(MelRecord):
-    """Quest record."""
-    classType = 'QUST'
-    _questFlags = Flags(0,Flags.getNames('startGameEnabled',None,'repeatedTopics','repeatedStages'))
-    stageFlags = Flags(0,Flags.getNames('complete'))
-    targetFlags = Flags(0,Flags.getNames('ignoresLocks'))
-
-    #--CDTA loader
-    class MelQustLoaders(DataDict):
-        """Since CDTA subrecords occur in three different places, we need
-        to replace ordinary 'loaders' dictionary with a 'dictionary' that will
-        return the correct element to handle the CDTA subrecord. 'Correct'
-        element is determined by which other subrecords have been encountered."""
-        def __init__(self,loaders,quest,stages,targets):
-            self.data = loaders
-            self.type_ctda = {'EDID':quest, 'INDX':stages, 'QSTA':targets}
-            self.ctda = quest #--Which ctda element loader to use next.
-        def __getitem__(self,key):
-            if key == 'CTDA': return self.ctda
-            self.ctda = self.type_ctda.get(key, self.ctda)
-            return self.data[key]
-
-    #--MelSet
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelFid('SCRI','script'),
-        MelString('FULL','full'),
-        MelString('ICON','iconPath'),
-        MelStruct('DATA','BB',(_questFlags,'questFlags',0),'priority'),
-        MelConditions(),
-        MelGroups('stages',
-            MelStruct('INDX','h','stage'),
-            MelGroups('entries',
-                MelStruct('QSDT','B',(stageFlags,'flags')),
-                MelConditions(),
-                MelString('CNAM','text'),
-                MelStruct('SCHR','4s4I',('unused1',null4),'numRefs','compiledSize','lastIndex','scriptType'),
-                MelBase('SCDA','compiled_p'),
-                MelString('SCTX','scriptText'),
-                MelScrxen('SCRV/SCRO','references')
-                ),
-            ),
-        MelGroups('targets',
-            MelStruct('QSTA','IB3s',(FID,'targetId'),(targetFlags,'flags'),('unused1',null3)),
-            MelConditions(),
-            ),
-        )
-    melSet.loaders = MelQustLoaders(melSet.loaders,*melSet.elements[5:8])
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreRace(MelRecord):
-    """Race record.
-
-    This record is complex to read and write. Relatively simple problems are the VNAM
-    which can be empty or zeroed depending on relationship between voices and
-    the fid for the race.
-
-    The face and body data is much more complicated, with the same subrecord types
-    mapping to different attributes depending on preceding flag subrecords (NAM0, NAM1,
-    NMAN, FNAM and INDX.) These are handled by using the MelRaceDistributor class
-    to dynamically reassign melSet.loaders[type] as the flag records are encountered.
-
-    It's a mess, but this is the shortest, clearest implementation that I could
-    think of."""
-
-    classType = 'RACE'
-    _flags = Flags(0L,Flags.getNames('playable'))
-
-    class MelRaceVoices(MelStruct):
-        """Set voices to zero, if equal race fid. If both are zero, then don't skip dump."""
-        def dumpData(self,record,out):
-            if record.maleVoice == record.fid: record.maleVoice = 0L
-            if record.femaleVoice == record.fid: record.femaleVoice = 0L
-            if (record.maleVoice,record.femaleVoice) != (0,0):
-                MelStruct.dumpData(self,record,out)
-
-    class MelRaceModel(MelGroup):
-        """Most face data, like a MelModel - MODT + ICON. Load is controlled by MelRaceDistributor."""
-        def __init__(self,attr,index):
-            MelGroup.__init__(self,attr,
-                MelString('MODL','modPath'),
-                MelBase('MODB','modb_p'),
-                MelBase('MODT','modt_p'),
-                MelString('ICON','iconPath'),)
-            self.index = index
-
-        def dumpData(self,record,out):
-            out.packSub('INDX','I',self.index)
-            MelGroup.dumpData(self,record,out)
-
-    class MelRaceIcon(MelString):
-        """Most body data plus eyes for face. Load is controlled by MelRaceDistributor."""
-        def __init__(self,attr,index):
-            MelString.__init__(self,'ICON',attr)
-            self.index = index
-        def dumpData(self,record,out):
-            out.packSub('INDX','I',self.index)
-            MelString.dumpData(self,record,out)
-
-    class MelRaceDistributor(MelNull):
-        """Handles NAM0, NAM1, MNAM, FMAN and INDX records. Distributes load
-        duties to other elements as needed."""
-        def __init__(self):
-            bodyAttrs = ('UpperBodyPath','LowerBodyPath','HandPath','FootPath','TailPath')
-            self.attrs = {
-                'MNAM':tuple('male'+text for text in bodyAttrs),
-                'FNAM':tuple('female'+text for text in bodyAttrs),
-                'NAM0':('head', 'maleEars', 'femaleEars', 'mouth',
-                'teethLower', 'teethUpper', 'tongue', 'leftEye', 'rightEye',)
-                }
-            self.tailModelAttrs = {'MNAM':'maleTailModel','FNAM':'femaleTailModel'}
-            self._debug = False
-
-        def getSlotsUsed(self):
-            return ('_loadAttrs',)
-
-        def getLoaders(self,loaders):
-            """Self as loader for structure types."""
-            for type in ('NAM0','MNAM','FNAM','INDX'):
-                loaders[type] = self
-
-        def setMelSet(self,melSet):
-            """Set parent melset. Need this so that can reassign loaders later."""
-            self.melSet = melSet
-            self.loaders = {}
-            for element in melSet.elements:
-                attr = element.__dict__.get('attr',None)
-                if attr: self.loaders[attr] = element
-
-        def loadData(self,record,ins,type,size,readId):
-            if type in ('NAM0','MNAM','FNAM'):
-                record._loadAttrs = self.attrs[type]
-                attr = self.tailModelAttrs.get(type)
-                if not attr: return
-            else: #--INDX
-                index, = ins.unpack('I',4,readId)
-                attr = record._loadAttrs[index]
-            element = self.loaders[attr]
-            for type in ('MODL', 'MODB', 'MODT', 'ICON'):
-                self.melSet.loaders[type] = element
-
-    #--Mel Set
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('FULL','full'),
-        MelString('DESC','text'),
-        MelFids('SPLO','spells'),
-        MelStructs('XNAM','Ii','relations',(FID,'faction'),'mod'),
-        MelStruct('DATA','14b2s4fI','skill1','skill1Boost','skill2','skill2Boost',
-                  'skill3','skill3Boost','skill4','skill4Boost','skill5','skill5Boost',
-                  'skill6','skill6Boost','skill7','skill7Boost',('unused1',null2),
-                  'maleHeight','femaleHeight','maleWeight','femaleWeight',(_flags,'flags',0L)),
-        MelRaceVoices('VNAM','2I',(FID,'maleVoice'),(FID,'femaleVoice')), #--0 same as race fid.
-        MelOptStruct('DNAM','2I',(FID,'defaultHairMale',0L),(FID,'defaultHairFemale',0L)), #--0=None
-        MelStruct('CNAM','B','defaultHairColor'), #--Int corresponding to GMST sHairColorNN
-        MelOptStruct('PNAM','f','mainClamp'),
-        MelOptStruct('UNAM','f','faceClamp'),
-        #--Male: Str,Int,Wil,Agi,Spd,End,Per,luck; Female Str,Int,...
-        MelStruct('ATTR','16B','maleStrength','maleIntelligence','maleWillpower',
-                  'maleAgility','maleSpeed','maleEndurance','malePersonality',
-                  'maleLuck','femaleStrength','femaleIntelligence',
-                  'femaleWillpower','femaleAgility','femaleSpeed',
-                  'femaleEndurance','femalePersonality','femaleLuck'),
-        #--Begin Indexed entries
-        MelBase('NAM0','_nam0',''), ####Face Data Marker, wbEmpty
-        MelRaceModel('head',0),
-        MelRaceModel('maleEars',1),
-        MelRaceModel('femaleEars',2),
-        MelRaceModel('mouth',3),
-        MelRaceModel('teethLower',4),
-        MelRaceModel('teethUpper',5),
-        MelRaceModel('tongue',6),
-        MelRaceModel('leftEye',7),
-        MelRaceModel('rightEye',8),
-        MelBase('NAM1','_nam1',''), ####Body Data Marker, wbEmpty
-        MelBase('MNAM','_mnam',''), ####Male Body Data Marker, wbEmpty
-        MelModel('maleTailModel'),
-        MelRaceIcon('maleUpperBodyPath',0),
-        MelRaceIcon('maleLowerBodyPath',1),
-        MelRaceIcon('maleHandPath',2),
-        MelRaceIcon('maleFootPath',3),
-        MelRaceIcon('maleTailPath',4),
-        MelBase('FNAM','_fnam',''), ####Female Body Data Marker, wbEmpty
-        MelModel('femaleTailModel'),
-        MelRaceIcon('femaleUpperBodyPath',0),
-        MelRaceIcon('femaleLowerBodyPath',1),
-        MelRaceIcon('femaleHandPath',2),
-        MelRaceIcon('femaleFootPath',3),
-        MelRaceIcon('femaleTailPath',4),
-        #--Normal Entries
-        MelFidList('HNAM','hairs'),
-        MelFidList('ENAM','eyes'),
-        MelBase('FGGS','fggs_p'), ####FaceGen Geometry-Symmetric
-        MelBase('FGGA','fgga_p'), ####FaceGen Geometry-Asymmetric
-        MelBase('FGTS','fgts_p'), ####FaceGen Texture-Symmetric
-        MelStruct('SNAM','2s',('snam_p',null2)),
-        #--Distributor for face and body entries.
-        MelRaceDistributor(),
-        )
-    melSet.elements[-1].setMelSet(melSet)
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreRefr(MelRecord):
-    classType = 'REFR'
-    _flags = Flags(0L,Flags.getNames('visible', 'canTravelTo'))
-    _parentFlags = Flags(0L,Flags.getNames('oppositeParent'))
-    _actFlags = Flags(0L,Flags.getNames('useDefault', 'activate','open','openByDefault'))
-    _lockFlags = Flags(0L,Flags.getNames(None, None, 'leveledLock'))
-    class MelRefrXloc(MelOptStruct):
-        """Handle older truncated XLOC for REFR subrecord."""
-        def loadData(self,record,ins,type,size,readId):
-            if size == 16:
-                MelStruct.loadData(self,record,ins,type,size,readId)
-                return
-            elif size == 12:
-                #--Else is skipping unused2
-                unpacked = ins.unpack('B3sIB3s',size,readId)
-            else:
-                raise ModError(ins.inName,u'Unexpected size encountered for REFR:XLOC subrecord: %i' % size)
-            unpacked = unpacked[:-2] + self.defaults[len(unpacked)-2:-2] + unpacked[-2:]
-            setter = record.__setattr__
-            for attr,value,action in zip(self.attrs,unpacked,self.actions):
-                if callable(action): value = action(value)
-                setter(attr,value)
-            if self._debug: print unpacked
-
-    class MelRefrXmrk(MelStruct):
-        """Handler for xmrk record. Conditionally loads next items."""
-        def loadData(self,record,ins,type,size,readId):
-            """Reads data from ins into record attribute."""
-            junk = ins.read(size,readId)
-            record.hasXmrk = True
-            insTell = ins.tell
-            insUnpack = ins.unpack
-            pos = insTell()
-            (type,size) = insUnpack('4sH',6,readId+'.FULL')
-            while type in ['FNAM','FULL','TNAM']:
-                if type == 'FNAM':
-                    value = insUnpack('B',size,readId)
-                    record.flags = MreRefr._flags(*value)
-                elif type == 'FULL':
-                    record.full = ins.readString(size,readId)
-                elif type == 'TNAM':
-                    record.markerType, record.unused5 = insUnpack('Bs',size,readId)
-                pos = insTell()
-                (type,size) = insUnpack('4sH',6,readId+'.FULL')
-            ins.seek(pos)
-            if self._debug: print ' ',record.flags,record.full,record.markerType
-
-        def dumpData(self,record,out):
-            if (record.flags,record.full,record.markerType,record.unused5) != self.defaults[1:]:
-                record.hasXmrk = True
-            if record.hasXmrk:
-                try:
-                    out.write(struct.pack('=4sH','XMRK',0))
-                    out.packSub('FNAM','B',record.flags.dump())
-                    value = record.full
-                    if value != None:
-                        out.packSub0('FULL',value)
-                    out.packSub('TNAM','Bs',record.markerType, record.unused5)
-                except struct.error:
-                    print self.subType,self.format,record.flags,record.full,record.markerType
-                    raise
-
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelFid('NAME','base'),
-        MelOptStruct('XTEL','I6f',(FID,'destinationFid'),'destinationPosX','destinationPosY',
-            'destinationPosZ','destinationRotX','destinationRotY','destinationRotZ'),
-        MelRefrXloc('XLOC','B3sI4sB3s','lockLevel',('unused1',null3),(FID,'lockKey'),('unused2',null4),(_lockFlags,'lockFlags'),('unused3',null3)),
-        MelOwnership(),
-        MelOptStruct('XESP','IB3s',(FID,'parent'),(_parentFlags,'parentFlags'),('unused4',null3)),
-        MelFid('XTRG','targetId'),
-        MelBase('XSED','seed_p'),
-        ####SpeedTree Seed, if it's a single byte then it's an offset into the list of seed values in the TREE record
-        ####if it's 4 byte it's the seed value directly.
-        MelOptStruct('XLOD','3f',('lod1',None),('lod2',None),('lod3',None)), ####Distant LOD Data, unknown
-        MelOptStruct('XCHG','f',('charge',None)),
-        MelOptStruct('XHLT','i',('health',None)),
-        MelXpci('XPCI'), ####fid, unknown
-        MelOptStruct('XLCM','i',('levelMod',None)),
-        MelFid('XRTM','xrtm'), ####unknown
-        MelOptStruct('XACT','I',(_actFlags,'actFlags',0L)), ####Action Flag
-        MelOptStruct('XCNT','i','count'),
-        MelRefrXmrk('XMRK','',('hasXmrk',False),(_flags,'flags',0L),'full','markerType',('unused5',null1)), ####Map Marker Start Marker, wbEmpty
-        MelBase('ONAM','onam_p'), ####Open by Default, wbEmpty
-        MelBase('XRGD','xrgd_p'),
-        MelOptStruct('XSCL','f',('scale',1.0)),
-        MelOptStruct('XSOL','B',('soul',None)), ####Was entirely missing. Confirmed by creating a test mod...it isn't present in any of the official esps
-        MelOptStruct('DATA','=6f',('posX',None),('posY',None),('posZ',None),('rotX',None),('rotY',None),('rotZ',None)),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreRegn(MelRecord):
-    """Region record."""
-    classType = 'REGN'
-    _flags = Flags(0L,Flags.getNames(
-        ( 2,'objects'),
-        ( 3,'weather'),
-        ( 4,'map'),
-        ( 6,'grass'),
-        ( 7,'sound'),))
-    obflags = Flags(0L,Flags.getNames(
-        ( 0,'conform'),
-        ( 1,'paintVertices'),
-        ( 2,'sizeVariance'),
-        ( 3,'deltaX'),
-        ( 4,'deltaY'),
-        ( 5,'deltaZ'),
-        ( 6,'Tree'),
-        ( 7,'hugeRock'),))
-    sdflags = Flags(0L,Flags.getNames(
-        ( 0,'pleasant'),
-        ( 1,'cloudy'),
-        ( 2,'rainy'),
-        ( 3,'snowy'),))
-
-    ####Lazy hacks to correctly read/write regn data
-    class MelRegnStructA(MelStructA):
-        """Handler for regn record. Conditionally dumps next items."""
-        def loadData(self,record,ins,type,size,readId):
-            if record.entryType == 2 and self.subType == 'RDOT':
-                MelStructA.loadData(self,record,ins,type,size,readId)
-            elif record.entryType == 3 and self.subType == 'RDWT':
-                MelStructA.loadData(self,record,ins,type,size,readId)
-            elif record.entryType == 6 and self.subType == 'RDGS':
-                MelStructA.loadData(self,record,ins,type,size,readId)
-            elif record.entryType == 7 and self.subType == 'RDSD':
-                MelStructA.loadData(self,record,ins,type,size,readId)
-
-        def dumpData(self,record,out):
-            """Conditionally dumps data."""
-            if record.entryType == 2 and self.subType == 'RDOT':
-                MelStructA.dumpData(self,record,out)
-            elif record.entryType == 3 and self.subType == 'RDWT':
-                MelStructA.dumpData(self,record,out)
-            elif record.entryType == 6 and self.subType == 'RDGS':
-                MelStructA.dumpData(self,record,out)
-            elif record.entryType == 7 and self.subType == 'RDSD':
-                MelStructA.dumpData(self,record,out)
-
-    class MelRegnString(MelString):
-        """Handler for regn record. Conditionally dumps next items."""
-        def loadData(self,record,ins,type,size,readId):
-            if record.entryType == 4 and self.subType == 'RDMP':
-                MelString.loadData(self,record,ins,type,size,readId)
-            elif record.entryType == 5 and self.subType == 'ICON':
-                MelString.loadData(self,record,ins,type,size,readId)
-
-        def dumpData(self,record,out):
-            """Conditionally dumps data."""
-            if record.entryType == 4 and self.subType == 'RDMP':
-                MelString.dumpData(self,record,out)
-            elif record.entryType == 5 and self.subType == 'ICON':
-                MelString.dumpData(self,record,out)
-
-    class MelRegnOptStruct(MelOptStruct):
-        """Handler for regn record. Conditionally dumps next items."""
-        def loadData(self,record,ins,type,size,readId):
-            if record.entryType == 7 and self.subType == 'RDMD':
-                MelOptStruct.loadData(self,record,ins,type,size,readId)
-
-        def dumpData(self,record,out):
-            """Conditionally dumps data."""
-            if record.entryType == 7 and self.subType == 'RDMD':
-                MelOptStruct.dumpData(self,record,out)
-
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('ICON','iconPath'),
-        MelStruct('RCLR','3Bs','mapRed','mapBlue','mapGreen',('unused1',null1)),
-        MelFid('WNAM','worldspace'),
-        MelGroups('areas',
-            MelStruct('RPLI','I','edgeFalloff'),
-            MelStructA('RPLD','2f','points','posX','posY')),
-        MelGroups('entries',
-            MelStruct('RDAT', 'I2B2s','entryType', (_flags,'flags'), 'priority', ('unused1',null2)), ####flags actually an enum...
-            MelRegnStructA('RDOT', 'IH2sf4B2H4s4f3H2s4s', 'objects', (FID,'objectId'), 'parentIndex',
-            ('unused1',null2), 'density', 'clustering', 'minSlope', 'maxSlope',
-            (obflags, 'flags'), 'radiusWRTParent', 'radius', ('unk1',null4),
-            'maxHeight', 'sink', 'sinkVar', 'sizeVar', 'angleVarX',
-            'angleVarY',  'angleVarZ', ('unused2',null2), ('unk2',null4)),
-            MelRegnString('RDMP', 'mapName'),
-## Disabled support due to bug when loading.
-## Apparently group records can't contain subrecords that are also present outside of the group.
-##            MelRegnString('ICON', 'iconPath'),  ####Obsolete? Only one record in oblivion.esm
-            MelRegnStructA('RDGS', 'I4s', 'grasses', (FID,'grass'), ('unk1',null4)),
-            MelRegnOptStruct('RDMD', 'I', ('musicType',None)),
-            MelRegnStructA('RDSD', '3I', 'sounds', (FID, 'sound'), (sdflags, 'flags'), 'chance'),
-            MelRegnStructA('RDWT', '2I', 'weather', (FID, 'weather'), 'chance')),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreRoad(MelRecord):
-    """Road structure. Part of large worldspaces."""
-    ####Could probably be loaded via MelStructA,
-    ####but little point since it is too complex to manipulate
-    classType = 'ROAD'
-    melSet = MelSet(
-        MelBase('PGRP','points_p'),
-        MelBase('PGRR','connections_p'),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-#------------------------------------------------------------------------------
-class MreSbsp(MelRecord):
-    """Subspace record."""
-    classType = 'SBSP'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelStruct('DNAM','3f','sizeX','sizeY','sizeZ'),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-#------------------------------------------------------------------------------
-class MreScpt(MelRecord):
-    """Script record."""
-    classType = 'SCPT'
-    _flags = Flags(0L,Flags.getNames('isLongOrShort'))
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelStruct('SCHR','4s4I',('unused1',null4),'numRefs','compiledSize','lastIndex','scriptType'),
-        #--Type: 0: Object, 1: Quest, 0x100: Magic Effect
-        MelBase('SCDA','compiled_p'),
-        MelString('SCTX','scriptText'),
-        MelGroups('vars',
-            MelStruct('SLSD','I12sB7s','index',('unused1',null4+null4+null4),(_flags,'flags',0L),('unused2',null4+null3)),
-            MelString('SCVR','name')),
-        MelScrxen('SCRV/SCRO','references'),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreSgst(MelRecord,MreHasEffects):
-    """Sigil stone record."""
-    classType = 'SGST'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelFull0(),
-        MelModel(),
-        MelString('ICON','iconPath'),
-        MelFid('SCRI','script'),
-        MelEffects(),
-        MelStruct('DATA','=BIf','uses','value','weight'),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreSkil(MelRecord):
-    """Skill record."""
-    classType = 'SKIL'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelStruct('INDX','i','skill'),
-        MelString('DESC','description'),
-        MelString('ICON','iconPath'),
-        MelStruct('DATA','2iI2f','action','attribute','specialization',('use0',1.0),'use1'),
-        MelString('ANAM','apprentice'),
-        MelString('JNAM','journeyman'),
-        MelString('ENAM','expert'),
-        MelString('MNAM','master'),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreSlgm(MelRecord):
-    """Soul gem record."""
-    classType = 'SLGM'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('FULL','full'),
-        MelModel(),
-        MelString('ICON','iconPath'),
-        MelFid('SCRI','script'),
-        MelStruct('DATA','If','value','weight'),
-        MelStruct('SOUL','B',('soul',0)),
-        MelStruct('SLCP','B',('capacity',1)),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreSoun(MelRecord):
-    """Sound record."""
-    classType = 'SOUN'
-    _flags = Flags(0L,Flags.getNames('randomFrequencyShift', 'playAtRandom',
-        'environmentIgnored', 'randomLocation', 'loop','menuSound', '2d', '360LFE'))
-    class MelSounSndd(MelStruct):
-        """SNDD is an older version of SNDX. Allow it to read in, but not set defaults or write."""
-        def loadData(self,record,ins,type,size,readId):
-            MelStruct.loadData(self,record,ins,type,size,readId)
-            record.staticAtten = 0
-            record.stopTime = 0
-            record.startTime = 0
-        def getSlotsUsed(self):
-            return ()
-        def setDefault(self,record): return
-        def dumpData(self,record,out): return
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('FNAM','soundFile'),
-        MelSounSndd('SNDD','=2BbsH2s','minDistance', 'maxDistance', 'freqAdjustment', ('unused1',null1),
-            (_flags,'flags'), ('unused2',null2)),
-        MelOptStruct('SNDX','=2BbsH2sH2B',('minDistance',None), ('maxDistance',None), ('freqAdjustment',None), ('unused1',null1),
-            (_flags,'flags',None), ('unused2',null2), ('staticAtten',None),('stopTime',None),('startTime',None),)
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreSpel(MelRecord,MreHasEffects):
-    """Spell record."""
-    classType = 'SPEL'
-    class SpellFlags(Flags):
-        """For SpellFlags, immuneSilence activates bits 1 AND 3."""
-        def __setitem__(self,index,value):
-            setter = Flags.__setitem__
-            setter(self,index,value)
-            if index == 1:
-                setter(self,3,value)
-    flags = SpellFlags(0L,Flags.getNames('noAutoCalc', 'immuneToSilence',
-        'startSpell', None,'ignoreLOS','scriptEffectAlwaysApplies','disallowAbsorbReflect','touchExplodesWOTarget'))
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelFull0(),
-        MelStruct('SPIT','3IB3s','spellType','cost','level',(flags,'flags',0L),('unused1',null3)),
-        # spellType = 0: Spell, 1: Disease, 3: Lesser Power, 4: Ability, 5: Poison
-        MelEffects(),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreStat(MelRecord):
-    """Static model record."""
-    classType = 'STAT'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelModel(),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreTree(MelRecord):
-    """Tree record."""
-    classType = 'TREE'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelModel(),
-        MelString('ICON','iconPath'),
-        MelStructA('SNAM','I','speedTree','seed'),
-        MelStruct('CNAM','5fi2f', 'curvature','minAngle','maxAngle',
-                  'branchDim','leafDim','shadowRadius','rockSpeed',
-                  'rustleSpeed'),
-        MelStruct('BNAM','2f','widthBill','heightBill'),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreWatr(MelRecord):
-    """Water record."""
-    classType = 'WATR'
-    _flags = Flags(0L,Flags.getNames('causesDmg','reflective'))
-    class MelWatrData(MelStruct):
-        """Handle older truncated DATA for WATR subrecord."""
-        def loadData(self,record,ins,type,size,readId):
-            if size == 102:
-                MelStruct.loadData(self,record,ins,type,size,readId)
-                return
-            elif size == 86:
-                #--Else 86 byte record (skips dispVelocity,
-                #-- dispFalloff, dispDampner, dispSize, and damage
-                #-- Two junk? bytes are tacked onto the end
-                #-- Hex editing and the CS confirms that it is NOT
-                #-- damage, so it is probably just filler
-                unpacked = ins.unpack('11f3Bs3Bs3BsB3s6f2s',size,readId)
-            elif size == 62:
-                #--Else 62 byte record (skips most everything
-                #-- Two junk? bytes are tacked onto the end
-                #-- No testing done, but assumed that its the same as the
-                #-- previous truncated record.
-                unpacked = ins.unpack('11f3Bs3Bs3BsB3s2s',size,readId)
-            elif size == 42:
-                #--Else 42 byte record (skips most everything
-                #-- Two junk? bytes are tacked onto the end
-                #-- No testing done, but assumed that its the same as the
-                #-- previous truncated record.
-                unpacked = ins.unpack('10f2s',size,readId)
-            elif size == 2:
-                #--Else 2 byte record (skips everything
-                #-- Two junk? bytes are tacked onto the end
-                #-- No testing done, but assumed that its the same as the
-                #-- previous truncated record.
-                unpacked = ins.unpack('2s',size,readId)
-            else:
-                raise ModError(ins.inName,_('Unexpected size encountered for WATR subrecord: %i') % size)
-            unpacked = unpacked[:-1]
-            unpacked += self.defaults[len(unpacked):]
-            setter = record.__setattr__
-            for attr,value,action in zip(self.attrs,unpacked,self.actions):
-                if callable(action): value = action(value)
-                setter(attr,value)
-            if self._debug: print unpacked
-
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('TNAM','texture'),
-        MelStruct('ANAM','B','opacity'),
-        MelStruct('FNAM','B',(_flags,'flags',0)),
-        MelString('MNAM','material'),
-        MelFid('SNAM','sound'),
-        MelWatrData('DATA', '11f3Bs3Bs3BsB3s10fH',('windVelocity',0.100),
-                    ('windDirection',90.0),('waveAmp',0.5),('waveFreq',1.0),('sunPower',50.0),
-                    ('reflectAmt',0.5),('fresnelAmt',0.0250),('xSpeed',0.0),('ySpeed',0.0),
-                    ('fogNear',27852.8),('fogFar',163840.0),('shallowRed',0),('shallowGreen',128),
-                    ('shallowBlue',128),('unused1',null1),('deepRed',0),('deepGreen',0),
-                    ('deepBlue',25),('unused2',null1),('reflRed',255),('reflGreen',255),
-                    ('reflBlue',255),('unused3',null1),('blend',50),('unused4',null3),('rainForce',0.1000),
-                    ('rainVelocity',0.6000),('rainFalloff',0.9850),('rainDampner',2.0000),
-                    ('rainSize',0.0100),('dispForce',0.4000),('dispVelocity', 0.6000),
-                    ('dispFalloff',0.9850),('dispDampner',10.0000),('dispSize',0.0500),('damage',0)),
-        MelFidList('GNAM','relatedWaters'),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreWeap(MelRecord):
-    """Weapon record."""
-    classType = 'WEAP'
-    _flags = Flags(0L,Flags.getNames('notNormalWeapon'))
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('FULL','full'),
-        MelModel(),
-        MelString('ICON','iconPath'),
-        MelFid('SCRI','script'),
-        MelFid('ENAM','enchantment'),
-        MelOptStruct('ANAM','H','enchantPoints'),
-        MelStruct('DATA','I2f3IfH','weaponType','speed','reach',(_flags,'flags',0L),
-            'value','health','weight','damage'),
-        #--weaponType = 0: Blade 1Hand, 1: Blade 2Hand, 2: Blunt 1Hand, 3: Blunt 2Hand, 4: Staff, 5: Bow
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreWrld(MelRecord):
-    """Worldspace record."""
-    classType = 'WRLD'
-    _flags = Flags(0L,Flags.getNames('smallWorld','noFastTravel','oblivionWorldspace',None,'noLODWater'))
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('FULL','full'),
-        MelFid('WNAM','parent'),
-        MelFid('CNAM','climate'),
-        MelFid('NAM2','water'),
-        MelString('ICON','mapPath'),
-        MelOptStruct('MNAM','2i4h',('dimX',None),('dimY',None),('NWCellX',None),('NWCellY',None),('SECellX',None),('SECellY',None)),
-        MelStruct('DATA','B',(_flags,'flags',0L)),
-        MelTuple('NAM0','ff','unknown0',(None,None)),
-        MelTuple('NAM9','ff','unknown9',(None,None)),
-        MelOptStruct('SNAM','I','sound'),
-        MelBase('OFST','ofst_p'),
-    )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#------------------------------------------------------------------------------
-class MreWthr(MelRecord):
-    """Weather record."""
-    classType = 'WTHR'
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelString('CNAM','lowerLayer'),
-        MelString('DNAM','upperLayer'),
-        MelModel(),
-        MelStructA('NAM0','3Bs3Bs3Bs3Bs','colors','riseRed','riseGreen','riseBlue',('unused1',null1),
-                   'dayRed','dayGreen','dayBlue',('unused2',null1),
-                   'setRed','setGreen','setBlue',('unused3',null1),
-                   'nightRed','nightGreen','nightBlue',('unused4',null1),
-                   ),
-        MelStruct('FNAM','4f','fogDayNear','fogDayFar','fogNightNear','fogNightFar'),
-        MelStruct('HNAM','14f',
-            'eyeAdaptSpeed', 'blurRadius', 'blurPasses', 'emissiveMult',
-            'targetLum', 'upperLumClamp', 'brightScale', 'brightClamp',
-            'lumRampNoTex', 'lumRampMin', 'lumRampMax', 'sunlightDimmer',
-            'grassDimmer', 'treeDimmer'),
-        MelStruct('DATA','15B',
-            'windSpeed','lowerCloudSpeed','upperCloudSpeed','transDelta',
-            'sunGlare','sunDamage','rainFadeIn','rainFadeOut','boltFadeIn',
-            'boltFadeOut','boltFrequency','weatherType','boltRed','boltBlue','boltGreen'),
-        MelStructs('SNAM','2I','sounds',(FID,'sound'),'type'),
-        )
-    __slots__ = MelRecord.__slots__ + melSet.getSlotsUsed()
-
-#-------------------------------------------------------------------------------
 
 #--Mergeable record types
 mergeClasses = (
